feat(driver-dom): add getDriver method
